refactor: split main.go into exporter.go and probes.go
main.go held the flags, the collector and the probes. It is now the flags
and main. exporter.go holds the collector, and probes.go holds the code
that reaches a target and reads its certificate.

The tests follow the same split. main_test.go becomes exporter_test.go
and probes_test.go, and TestDaysLeft moves next to daysLeft.
