+++ collect_test.go
@@ -155,42 +155,3 @@ func TestTargetLeavesTheOutput(t *testing.T) {
 		t.Errorf("families: got %v, want none", names)
 	}
 }
-
-func TestDaysLeft(t *testing.T) {
-	now := time.Date(2026, 8, 6, 12, 0, 0, 0, time.UTC)
-
-	tests := []struct {
-		name     string
-		notAfter time.Time
-		want     float64
-	}{
-		{
-			name:     "one day left",
-			notAfter: now.Add(24 * time.Hour),
-			want:     1,
-		},
-		{
-			name:     "half a day left",
-			notAfter: now.Add(12 * time.Hour),
-			want:     0.5,
-		},
-		{
-			name:     "the certificate ended a day ago",
-			notAfter: now.Add(-24 * time.Hour),
-			want:     -1,
-		},
-		{
-			name:     "the certificate ends now",
-			notAfter: now,
-			want:     0,
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			if got := daysLeft(test.notAfter, now); got != test.want {
-				t.Errorf("daysLeft: got %v, want %v", got, test.want)
-			}
-		})
-	}
-}
